pyproject.toml
[tool.poetry]
name = "anime-diffusion"
version = "0.1.0"
description = "Basic implementation of a diffusion model to generate anime faces."
authors = ["Pierrot LeCon <pstmrtem@outlook.com>"]
readme = "README.md"
packages = [{include = "anime_diffusion"}]

[tool.poetry.dependencies]
python = "^3.10"
kaggle = "^1.5.13"
kornia = "^0.6.12"
torchinfo = "^1.8.0"
wandb = "^0.15.3"
tqdm = "^4.65.0"
torchvision = "^0.15.2"
positional-encodings = {extras = ["pytorch"], version = "^6.0.1"}
hydra-core = "^1.3.2"
einops = "^0.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
